subProj 3 UDS Diagnostic Tester/requirements.txt
pytest==8.0.0
pytest-html==4.1.1
python-can==4.3.1
udsoncan==1.25.2
can-isotp==2.0.4
